% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcctzuser-package.R
\docType{package}
\name{rcctzuser-package}
\alias{rcctzuser}
\alias{rcctzuser-package}
\title{rcctzuser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/rcctzuser}
  \item Report bugs at \url{https://github.com/DavisVaughan/rcctzuser/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
